Refactored some stuff ..
